+++ plugins/@vcmap-show-case/category-tester/src/CollectionComponentOptions.vue
@@ -24,13 +24,22 @@
         />
       </v-col>
     </v-row>
+    <v-row no-gutters>
+      <v-col>
+        <VcsLabel>{{ $st(`pagination`) }}</VcsLabel>
+      </v-col>
+      <v-col>
+        <VcsCheckbox v-model="pagination" />
+      </v-col>
+    </v-row>
   </v-container>
 </template>
 
 <script>
-  import { inject, isRef, reactive } from 'vue';
+  import { inject, isRef, reactive, watch, ref } from 'vue';
   import { VContainer, VRow, VCol } from 'vuetify/components';
   import { VcsCheckbox, VcsLabel, VcsTextField } from '@vcmap/ui';
+  import { Collection } from '@vcmap/core';
 
   /**
    * Shows CollectionComponentOptions of a provided collectionComponent
@@ -45,7 +54,9 @@
       VRow,
       VCol,
     },
+    props: {},
     setup() {
+      /** @type {import("@vcmap/ui").CollectionComponentClass} */
       const collectionComponent = inject('collectionComponent');
       const localOptions = reactive({
         draggable: collectionComponent.draggable,
@@ -54,7 +65,34 @@
         renamable: collectionComponent.renamable,
         removable: collectionComponent.removable,
       });
-
+      // clone is for demonstration purposes only: caches the original items while pagination is active.
+      const collectionClone = new Collection();
+      const pagination = ref(!!collectionComponent.pagination.value);
+      watch(pagination, (value) => {
+        if (value) {
+          collectionClone.clear();
+          [...collectionComponent.collection].forEach((item) => {
+            collectionClone.add(item);
+          });
+          collectionComponent.setPagination({
+            // just for demonstration purposes -> get Items should fetch items from server.
+            getItems(startIndex, count) {
+              return {
+                items: [...collectionClone].slice(
+                  startIndex,
+                  count + startIndex,
+                ),
+                total: collectionClone.size,
+              };
+            },
+          });
+        } else {
+          collectionComponent.setPagination(undefined);
+          [...collectionClone].forEach((item) => {
+            collectionComponent.collection.add(item);
+          });
+        }
+      });
       return {
         title: collectionComponent.title,
         keys: Object.keys(localOptions),
@@ -67,6 +105,7 @@
             collectionComponent[key] = value;
           }
         },
+        pagination,
       };
     },
   };